having a type error conflict between nan and uint32. Force nan to uint32 (which makes it zero... which is how it was originally before I thought nan was a neat way around. I could just change all indexes back to double precision ...).

+++ src/commonGrouping.m
@@ -20,6 +20,6 @@ function [Gc,ix] = commonGrouping(G)
 end
 
 function N = cellmax(C)
-    C(cellfun('isempty',C)) = {nan}; %handle empty cells
+    C(cellfun('isempty',C)) = {uint32(nan)}; %handle empty cells
     N = uint32(max(cellfun(@max, C),[],1,'omitnan'));
 end
